Handle metadata & extra Stripe params
Instead of constructing the charge params to send to Stripe from the minimum required components of the charge params provided to createSplitTenderCharge, copy those params and remove the Lightrail-specific keys. This ensures that using this library will not break an existing checkout flow which may require extra information (that we shouldn't try to predict/control) to be passed to Stripe.

Also add split tender transaction details to the metadata that is sent to Stripe (this is important for traceability).

Note: params are being shallow copied.
Note: Lightrail-specific keys may change and need to be revisited.