Rewrite `createSplitTenderCharge` to take Stripe object OR key
Instead of having a wrapper method for the sole purpose of accepting a Stripe secret key as the Stripe auth parameter (instead of an configured Stripe isntance), allow the default method to accept either one. Relies on a simple helper method to instantiate a new Stripe object if a secret key is passed in instead.