Re-throw Stripe error from createSplitTenderCharge
If the Stripe part of a split tender transaction throws an error, it should be surfaced. The previous flow swallowed it by catching the Stripe error, voiding the Lightrail transaction, and returning the 'split tender charge' object containing only a voided Lightrail transaction. New flow catches, voids, and then re-throws the Stripe error.