Add test for Lightrail errors; group error tests together
Adds a single test to check that Lightrail transaction errors are thrown, not swallowed. Until mocks/stubs are implemented, this can only check for errors in the first part of the Lightrail transaction (post a pending transaction): another test should be added later to cover errors that are thrown during the void/capture phase.